libdpkg: Cast strlen() return value to ssize_t to match write() return type
Warned-by: clang

+++ lib/dpkg/t/t-buffer.c
@@ -62,7 +62,7 @@ test_fdio_hash(void)
 	test_pass(fd_md5(fd, hash, -1, NULL) >= 0);
 	test_str(hash, ==, ref_hash_empty);
 
-	test_pass(write(fd, str_test, strlen(str_test)) == strlen(str_test));
+	test_pass(write(fd, str_test, strlen(str_test)) == (ssize_t)strlen(str_test));
 	test_pass(lseek(fd, 0, SEEK_SET) == 0);
 
 	test_pass(fd_md5(fd, hash, -1, NULL) >= 0);
